bump 0.3.0(steam-web-api): Error types for steam_web_api,
with eresult/e-error_message header checking and fixes to POST endpoints.

Post endpoints previously were not working, because they needed to be sent as
form urlencoded. A minor change on web-api-derive to allow referencing the
endpoint parameters and then serializing them with reqwest form.
A minor change also on macros by matching the http methods was needed, since
only POST required this change.

Now the API also checks for the X-eresult and X-error_message for error
information. Important endpoints impacted by this are DeclineTradeOffer
and CancelTradeOffer, because a cancel request for a already accepted trade
offer could only be found with X-eresult error header.

The EResult error can be matched with the EResult enum, for more precise error
handling.

Signed-off-by: Martin <martin@hotmail.com.br>