(auth): deps bump, some changes to inner reqwest client, clippy lints
Policy of redirection was increase from none to 5 to acommodate changes to
`steam-trading`. `steam-trading` needed redirection to check if user receiving
the trade has Steam Guard enabled through the `marketEligibility/..` route.

If this causes any error, there is the need to set those cookies directly on
`steam-trading`, check if the redirection is still needed, and if yes, we should
return the policy of redirection back to none;

Signed-off-by: Martin <martin@hotmail.com.br>